Use rally-openstack's upper constraints
docker-rally images built at present without upper constraints fail at
the end of a rally/tempest verification with the following errors:

    <class 'memoryview'> can't be decoded
    Missing argument: --uuid

It's unclear what is causing the problem, but rebuilding the image with
dependencies pinned to the last known working image (the 1.4-dev tag)
results in a working image.

Looking at the rally-openstack source, the project maintains its own
upper constraints file that it uses in testing. It makes sense for us to
also use this to aim for a consistent and tested set of dependencies.

This change also adds Docker build args for the rally-openstack repo,
version and upper constraints.